ZIP File
Created on a work server, this app is used to interpret medical documents. The challenge with medicaid authorizations is that the patterns for the dates they cover can be irregular (i.e. they are not always simply six months long, or one month long.), which means the allotted units are not always the same per week. This app makes calculating simple by simply taking the start and end dates and figuring out how many actual days are covered, and then taking the units and breaking them out by weekly amounts and term amounts (Term being the duration of the authorization). It even has a build in section pre-filled with standard notes commonly used when updating multiple authorizations. This calculator converts 1 unit into the standard 15 minute increments equal to 4 units per hour.